==> scripts/00-data-setup.Rmd <==
---
title: "AEIS Data Example -- Generating a Synthetic RCT"
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
      out_dir <- "compiled_tutorials";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
output: github_document
always_allow_html: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(optmatch)
library(dplyr)
library(tidyr)
library(kableExtra)
```

## Load Data

```{r}
load("../input/MS_data_public.Rdata")
var.names.clean <- read.csv("../input/var_names.csv")
```

## Clean some variable names for interpretability

```{r}
covs_ms <- covs_ms %>%
  rename(all_stud_n = CPETALLC_67, grade8_n = CPETG08C_67, stud_teach_rat = CPSTKIDR_67,
         all_exp = CPFPAALLT_67, inst_exp = CPFEAINSP_67, lead_exp = CPFEAADSP_67, supp_exp = CPFEASUPP_67,
         ed_aide = CPSETOFC_67, teach_salary = CPSTTOSA_67, 
         teach_expr = CPSTEXPA_67, perc_teach_female = CPSTFEFP_67,
         perc_teach_white = CPSTWHFP_67, perc_teach_black = CPSTBLFP_67, perc_teach_hisp = CPSTHIFP_67,
         perc_stud_black = CPETBLAP_67, perc_stud_hisp = CPETHISP_67, 
         perc_stud_api = CPETPACP_67, perc_stud_white = CPETWHIP_67,
         perc_stud_alp = CPETSPEP_67, perc_stud_bil = CPETBILP_67, perc_stud_tag = CPETGIFP_67) %>%
  select(CAMPUS, GRDSPAN, starts_with("pre"),all_of(var.names.clean$var_clean), everything())

covs_ms_noprep <- covs_ms_noprep %>%
  rename(all_stud_n = CPETALLC_67, grade8_n = CPETG08C_67, stud_teach_rat = CPSTKIDR_67,
         all_exp = CPFPAALLT_67, inst_exp = CPFEAINSP_67, lead_exp = CPFEAADSP_67, supp_exp = CPFEASUPP_67,
         ed_aide = CPSETOFC_67, teach_salary = CPSTTOSA_67, 
         teach_expr = CPSTEXPA_67, perc_teach_female = CPSTFEFP_67,
         perc_teach_white = CPSTWHFP_67, perc_teach_black = CPSTBLFP_67, perc_teach_hisp = CPSTHIFP_67,
         perc_stud_black = CPETBLAP_67, perc_stud_hisp = CPETHISP_67, 
         perc_stud_api = CPETPACP_67, perc_stud_white = CPETWHIP_67,
         perc_stud_alp = CPETSPEP_67, perc_stud_bil = CPETBILP_67, perc_stud_tag = CPETGIFP_67) %>%
  select(CAMPUS, GRDSPAN, starts_with("pre"),all_of(var.names.clean$var_clean), everything())
```

Here are descriptions of the variables that we will use in matching:

```{r, echo = F}
var.names.clean %>%
  kable(col.names = c("Original Variable", "Variable Description", "Clean Name"),
                     align = "lll") %>% 
  row_spec(row = 0, align = "c", bold=T) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"))
```

## Form Fake/Example Paired RCT

First, match pairs of schools on baseline characteristics. Note, that this is not how one would design a paired trial because I used a bipartite pair matching algorithm from `optmatch` (which is just the package that I know how to use).

This results in 146 pairs of schools based on their previous 8th grade Math TAKS performance, size of the school, student and teacher demographics, and total school funding.

```{r}
set.seed(25)

poss_rct <- covs_ms_noprep %>%
  filter(!is.na(grade8_n)) %>%
  select(CAMPUS) %>%
  left_join(covs_ms)

# create a random grouping to be able to do bipartite matching
poss_rct$Z <- sample(0:1, nrow(poss_rct), replace = T)

# calculate distances between schools
dist_mat <- match_on(Z ~ premA + premB + premF + all_stud_n + stud_teach_rat + all_exp + 
                       perc_teach_white + perc_stud_hisp + 
                       perc_stud_alp + perc_stud_bil, data = poss_rct, caliper = 1)
summary(dist_mat)

# exclude distances outside of the caliper
initialmatch <- fullmatch(dist_mat, data = poss_rct)
excl <- which(is.na(initialmatch))

# and recalculate distances
dist.update <- match_on(Z ~ premA + premB + premF + all_stud_n + stud_teach_rat + all_exp + 
                       perc_teach_white + perc_stud_hisp + 
                       perc_stud_alp + perc_stud_bil, dat = poss_rct[-excl,]) 

# final optimal pair matching
fakematch <- pairmatch(dist.update,  data = poss_rct)
summary(fakematch)
```

Select 25 random pairs from those matched to make up our paired trial.

```{r}
set.seed(125)

match.ids <- unique(fakematch[which(!is.na(fakematch))])
rct.ids <- sample(match.ids, 25)
```

Randomly assign one school within each pair to receive treatment `Tr` and save RCT information. We also include a Bernoulli randomized treatment assignment `TrBern` (ignoring pairs).

```{r}
set.seed(73)

rct <- poss_rct %>%
  select(CAMPUS) %>%
  mutate(match = fakematch) %>%
  filter(match %in% rct.ids) %>%
  mutate(match = factor(match,levels = rct.ids, labels = LETTERS[1:25])) %>%
  group_by(match) %>% 
  mutate(k = row_number()) %>%
  pivot_wider(values_from = CAMPUS, names_from = k, names_prefix = "CAMPUS") %>%
  mutate(Tr1 = sample(0:1,1),
         Tr2 = 1-Tr1) %>%
  pivot_longer(!match,
               names_to = c(".value","k"),
               names_pattern = "(.*)(1|2)") %>%
  select(CAMPUS, match, k, Tr) %>%
  ungroup() %>%
  arrange(match, k) %>%
  left_join(select(covs_ms_noprep, CAMPUS, clust_size = CPETG08C_78))

rct$TrBern = sample(0:1, 50, replace = T)
```

```{r}
save(rct, file = "../data/rct_schools.Rdata")
```


Example Paired RCT schools:

```{r, echo = F}
rct %>%
  select(-k) %>%
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "hover","condensed"))
```

## Rename Outcomes for Clarity

The outcome data consists of variables representing the **8th grade math TAKS passing rates** for different subgroups of students and two different school years. The variables follow the naming structure `outm[A][08]` where `[A]` takes the values of A, B, H, W, M, F, and E which indicate subgroups as in the table below and `[08]` takes the value of 08 or 09, indicating the 2007/8 or 2008/9 school years.

| Code |   Subgroup   |
|------|--------------|
|   A  | All Students |
|   B  | Black        |
|   E  | Low SES      |
|   F  | Female       |
|   H  | Hispanic     |
|   M  | Male         |
|   W  | White        |

For this tutorial, we will focus on the overall `A`, 2008 `08` math TAKS passing rate, so we rename that column `taks08`, but leave the remaining variable names as-is, if you would like to explore.

```{r}
out_ms <- out_ms %>%
  rename(taks08 = outmA08) %>%
  select(CAMPUS, taks08, everything())
```



## Save Cleaned Data

```{r}
rct_dat = rct %>%
  left_join(out_ms, by = "CAMPUS") %>%
  left_join(covs_ms, by = "CAMPUS") %>%
  select(CAMPUS, GRDSPAN, match, k , Tr, TrBern, taks08, starts_with("pre"), 
         all_of(var.names.clean$var_clean), everything())

aux_dat = out_ms %>%
  filter(!(CAMPUS %in% rct$CAMPUS)) %>%
  left_join(covs_ms, by = "CAMPUS") %>%
  select(CAMPUS, GRDSPAN, taks08, starts_with("pre"), 
         all_of(var.names.clean$var_clean), everything())

save(rct_dat, aux_dat, covs_ms_noprep, file = "../data/rct_aux_dat.Rdata")
```

```{r}
#subset to specific variable names for Shiny app demo 
subset_var_names <- read.csv("../input/subset_var_names.csv", col.names = "col_name")

aux_dat_small <- aux_dat %>%
  select(subset_var_names$col_name)

write.csv(aux_dat_small, "../data/aux_dat_small.csv", row.names = FALSE)
```
